- Did some cleanup. - Did some refactoring. - BugFix: Home page stops at display loading spinner for visitors.
